+++ certification/index.ts
@@ -1,0 +1,2 @@
+export * from "./gai08c_certification";
+export * from "./run_independent_certification";